requirements.txt
pyyaml
gym
pandas==1.3.0
matplotlib==3.4.2
numpy==1.21.1
torch
jupyterlab==3.1.1
isort==5.10.1
black==22.1.0